✅ test: extract progress regression test into its own suite (#72)
ModelManagerTests was over swiftlint's 400-line file_length limit after
the throttle regression test was added. Move the new test, its mock
(`SubOneProgressMockDownloader`), and the `StateSnapshots` observation
helper into a dedicated `ModelManagerProgressTests` suite.

Pin `expectedFileSize: 0` explicitly in the new suite — it constructs
`ModelManager` directly rather than via the existing `makeSUT` helper,
and the production default (3.1 GB) would otherwise reject the 1000-byte
mock payload during integrity verification.

Pre-clean Application Support / Caches files at the start of the test:
Swift Testing runs suites in parallel by default, and `.serialized` only
orders tests *within* a suite — `ModelManagerTests` and the new suite
share filesystem paths.
